+++ src/models/Settings.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Read Time plugin for Craft CMS 4.x
+ * Read Time plugin for Craft CMS 5.x
  *
  * Calculate the estimated read time for content.
  *
@@ -8,25 +8,24 @@
  * @copyright Copyright (c) 2018 Jalen Davenport
  */
 
+declare(strict_types=1);
+
 namespace jalendport\readtime\models;
 
 use craft\base\Model;
 
 class Settings extends Model
 {
-    // Public Properties
-    // =========================================================================
-
+    /**
+     * @var int Average reading speed, in words per minute.
+     */
     public int $wordsPerMinute = 200;
-
-    // Public Methods
-    // =========================================================================
 
     public function rules(): array
     {
         return [
             [['wordsPerMinute'], 'required'],
-            [['wordsPerMinute'], 'number', 'integerOnly' => true]
+            [['wordsPerMinute'], 'number', 'integerOnly' => true],
         ];
     }
 }